requirements.txt
torch==2.3.1
transformers==4.45.2
datasets==2.20.0
accelerate==0.26.1
jupyter==1.0.0
numpy==1.26.3
pillow==10.2.0
torch==2.3.1
peft==0.9.0
diffusers==0.30.3
timm==0.9.16
gradio
spaces
fastapi
uvicorn
pillow
pydantic
transformers
scipy
safetensors
python-socketio[asyncio_server]
fastapi-socketio
accelerate

